[backend] use fewer headers in allocator
allocator does not need to include vulkan.hpp indirectly - this has
dramatic effect on its compilation time
