+++ modules/le_backend_vk/le_backend_types_internal.h
@@ -13,33 +13,6 @@ constexpr uint8_t VK_MAX_BOUND_DESCRIPTOR_SETS = 8;
 constexpr uint8_t VK_MAX_COLOR_ATTACHMENTS     = 16; // maximum number of color attachments to a renderpass
 
 // ----------------------------------------------------------------------
-// Preprocessor Macro utilities
-//
-
-// Wraps an enum of `enum_name` in a struct with `struct_name` so
-// that it can be opaquely passed around, then unwrapped.
-#define LE_WRAP_ENUM_IN_STRUCT( enum_name, struct_name ) \
-	struct struct_name {                                 \
-		enum_name data;                                  \
-		          operator const enum_name&() const {    \
-            return data;                       \
-		}                                                \
-		operator enum_name&() {                          \
-			return data;                                 \
-		}                                                \
-	}
-
-// ----------------------------------------------------------------------
-// Utility methods
-//
-template <typename T>
-static constexpr typename std::underlying_type<T>::type enumToNum( const T& enumVal ) {
-	return static_cast<typename std::underlying_type<T>::type>( enumVal );
-};
-
-// ----------------------------------------------------------------------
-
-LE_WRAP_ENUM_IN_STRUCT( vk::Format, VkFormatEnum ); // define wrapper struct `VkFormatEnum`
 
 // ----------------------------------------------------------------------
 
@@ -217,8 +190,8 @@ struct LeRenderPass {
 
 	LeRenderPassType type;
 
-	vk::Framebuffer         framebuffer;
-	vk::RenderPass          renderPass;
+	VkFramebuffer           framebuffer;
+	VkRenderPass            renderPass;
 	uint32_t                width;
 	uint32_t                height;
 	vk::SampleCountFlagBits sampleCount;    // We store this with renderpass, as sampleCount must be same for all color/depth attachments
